--- GNNVPR/collection/minw.csv
@@ -1,79 +1,0 @@
-architecture,circuit,minw
-EArch,alu4,76
-EArch,apex2,104
-EArch,apex4,88
-EArch,bigkey,44
-EArch,clma,126
-EArch,des,60
-EArch,diffeq,72
-EArch,dsip,46
-EArch,elliptic,102
-EArch,ex1010,106
-EArch,ex5p,110
-EArch,frisc,116
-EArch,misex3,86
-EArch,pdc,140
-EArch,s298,58
-EArch,s38417,70
-EArch,s38584,78
-EArch,seq,98
-EArch,spla,122
-EArch,tseng,68
-stratixiv_arch,alu4,50
-stratixiv_arch,apex2,66
-stratixiv_arch,apex4,64
-stratixiv_arch,bigkey,34
-stratixiv_arch,clma,106
-stratixiv_arch,des,42
-stratixiv_arch,diffeq,50
-stratixiv_arch,dsip,32
-stratixiv_arch,elliptic,82
-stratixiv_arch,ex1010,82
-stratixiv_arch,ex5p,66
-stratixiv_arch,frisc,96
-stratixiv_arch,misex3,60
-stratixiv_arch,pdc,112
-stratixiv_arch,s298,42
-stratixiv_arch,s38417,72
-stratixiv_arch,s38584.1,64
-stratixiv_arch,seq,66
-stratixiv_arch,spla,96
-stratixiv_arch,tseng,52
-stratixiv_arch,CHERI_stratixiv_arch_timing,262
-stratixiv_arch,CH_DFSIN_stratixiv_arch_timing,192
-stratixiv_arch,EKF-SLAM_Jacobians_stratixiv_arch_timing,270
-stratixiv_arch,JPEG_stratixiv_arch_timing,228
-stratixiv_arch,MCML_stratixiv_arch_timing,152
-stratixiv_arch,MMM_stratixiv_arch_timing,120
-stratixiv_arch,Reed_Solomon_stratixiv_arch_timing,96
-stratixiv_arch,SURF_desc_stratixiv_arch_timing,160
-stratixiv_arch,carpat_stratixiv_arch_timing,150
-stratixiv_arch,fir_cascade_stratixiv_arch_timing,88
-stratixiv_arch,jacobi_stratixiv_arch_timing,140
-stratixiv_arch,leon2_stratixiv_arch_timing,174
-stratixiv_arch,leon3mp_stratixiv_arch_timing,188
-stratixiv_arch,radar20_stratixiv_arch_timing,102
-stratixiv_arch,random_stratixiv_arch_timing,136
-stratixiv_arch,smithwaterman_stratixiv_arch_timing,140
-stratixiv_arch,stap_steering_stratixiv_arch_timing,112
-stratixiv_arch,sudoku_check_stratixiv_arch_timing,234
-stratixiv_arch,ucsb_152_tap_fir_stratixiv_arch_timing,104
-stratixiv_arch,uoft_raytracer_stratixiv_arch_timing,228
-stratixiv_arch,wb_conmax_stratixiv_arch_timing,110
-stratixiv_arch,SLAM_spheric_stratixiv_arch_timing,236
-stratixiv_arch,bitcoin_miner_stratixiv_arch_timing,36
-stratixiv_arch,bitonic_mesh_stratixiv_arch_timing,182
-stratixiv_arch,cholesky_bdti_stratixiv_arch_timing,176
-stratixiv_arch,cholesky_mc_stratixiv_arch_timing,172
-stratixiv_arch,dart_stratixiv_arch_timing,244
-stratixiv_arch,des90_stratixiv_arch_timing,220
-stratixiv_arch,gsm_switch_stratixiv_arch_timing,200
-stratixiv_arch,mes_noc_stratixiv_arch_timing,190
-stratixiv_arch,minres_stratixiv_arch_timing,176
-stratixiv_arch,neuron_stratixiv_arch_timing,124
-stratixiv_arch,openCV_stratixiv_arch_timing,192
-stratixiv_arch,segmentation_stratixiv_arch_timing,170
-stratixiv_arch,sparcT1_core_stratixiv_arch_timing,270
-stratixiv_arch,sparcT2_core_stratixiv_arch_timing,268
-stratixiv_arch,stap_qrd_stratixiv_arch_timing,174
-stratixiv_arch,stereo_vision_stratixiv_arch_timing,122
