--- GNNVPR/evaluation/metrics/metric.py
@@ -1,3 +1,0 @@
-"""This will contain common utility functions used
-    by the report notebook for generating ables & graphs
-"""
