--- GNNVPR/main.py
@@ -1,3 +1,0 @@
-"""This will run the entire model framework start to
-finish given a selected set of benchmarks.
-"""
