--- GNNVPR/runner.py
@@ -1,5 +1,0 @@
-"""Utility Functions to batch run sections.
-"""
-def runner():
-    # We run the entire framework
-
