--- output/models/stub.txt
@@ -1,1 +1,0 @@
-This is where sample & output models go.
