--- tests/test_gpu.py
@@ -1,2 +1,0 @@
-"""This Test Suite is used as a utilty to set the device for pytorch.
-"""
