--- tests/test_outputs.py
@@ -1,17 +1,0 @@
-"""This Test Suite Ensures the outputs from step to step are:
-1. Not Empty
-2. Rightly Formatted
-"""
-
-import unittest
-# from src.metrics import metric
-
-
-class TestMetric(unittest.TestCase):
-
-    def test_metric(self):
-        self.assertEqual(0, 0)
-
-
-if __name__ == "__main__":
-    unittest.main()
